waikato_aerial/similarity/requirements.txt
datasets==3.2.0
click==8.1.8
pillow==11.1.0
numpy==2.2.2
torch==2.6.0
torcheval==0.0.7
torchvision==0.21.0
transformers==4.48.3
tqdm==4.67.1
accelerate==1.3.0
pillow==11.1.0
absl-py==2.1.0